SmokeTests/SCHEDULING/DSACore/requirements.txt
nose
nose-testconfig
urllib3
